type 에러 해결, 미세먼지 및 바다날씨 웹 페이지 웹뷰에 추가함, 미세먼지와 바다날씨는 수정이 필요함

+++ lib/weather_information/screens/weather_screen.dart
@@ -35,7 +35,7 @@ class _WeatherScreenState extends State<WeatherScreen> {
   }
 
   void updateData(dynamic weatherData, dynamic airData) {
-    double temp2 = weatherData['main']['temp'];
+    double temp2 = weatherData['main']['temp'].toDouble();
     int condition = weatherData['weather'][0]['id'];
     int index = airData['list'][0]['main']['aqi'];
     des = weatherData['weather'][0]['description'];
